Moving main actions to a new topBar instead of 3 dots menu
