+++ packages/bruno-app/src/components/Sidebar/index.js
@@ -15,7 +15,6 @@ const MAX_LEFT_SIDEBAR_WIDTH = 600;
 
 const Sidebar = () => {
   const leftSidebarWidth = useSelector((state) => state.app.leftSidebarWidth);
-  const preferencesOpen = useSelector((state) => state.app.showPreferences);
 
   const [asideWidth, setAsideWidth] = useState(leftSidebarWidth);
 
@@ -78,8 +77,6 @@ const Sidebar = () => {
     <StyledWrapper className="flex relative h-screen">
       <aside>
         <div className="flex flex-row h-screen w-full">
-          {preferencesOpen && <Preferences onClose={() => dispatch(showPreferences(false))} />}
-
           <div className="flex flex-col w-full" style={{ width: asideWidth }}>
             <div className="flex flex-col flex-grow">
               <TitleBar />
@@ -87,14 +84,6 @@ const Sidebar = () => {
             </div>
 
             <div className="footer flex px-1 py-2 absolute bottom-0 left-0 right-0 items-center cursor-pointer select-none">
-              <div className="flex items-center ml-1 text-xs ">
-                <IconSettings
-                  size={18}
-                  strokeWidth={1.5}
-                  className="mr-2  hover:text-gray-700"
-                  onClick={() => dispatch(showPreferences(true))}
-                />
-              </div>
               <div className="pl-1" style={{ position: 'relative', top: '3px' }}>
                 <GitHubButton
                   href="https://github.com/usebruno/bruno"
